♻️ Refactor: Centralize NL2Agent MCP tool definitions
