+++ backend/tool_collection/mcp/local_mcp_service.py
@@ -1,7 +1,10 @@
 from fastmcp import FastMCP
 
 from tool_collection.mcp.nl2agent_mcp_tools import (
+    NL2AGENT_MCP_TOOL_META,
+    NL2A_WRAPPER_DESCRIPTION,
     NL2A_WRAPPER_NAME,
+    SEARCH_INSTALLED_MCP_TOOLS_DESCRIPTION,
     SEARCH_INSTALLED_MCP_TOOLS_NAME,
     nl2a_wrapper as _nl2a_wrapper,
     search_installed_mcp_tools as _search_installed_mcp_tools,
@@ -14,36 +17,18 @@ LOCAL_MCP_TOOL_NAME_OVERRIDES = {
 
 # Create MCP server
 local_mcp_service = FastMCP("local")
-_search_installed_mcp_tool = local_mcp_service.tool(
+local_mcp_service.tool(
     _search_installed_mcp_tools,
     name=SEARCH_INSTALLED_MCP_TOOLS_NAME,
-    description=(
-        "Search the current tenant's installed and available MCP tools using keywords. "
-        "Returns a structured JSON observation ordered by relevance. "
-        "Call the tool as `result = search_installed_mcp_tools(...)`, then use "
-        "`print(result)` to preserve the returned JSON unchanged in execution logs."
-    ),
-    meta={"nexent_internal": True},
+    description=SEARCH_INSTALLED_MCP_TOOLS_DESCRIPTION,
+    meta=NL2AGENT_MCP_TOOL_META,
 )
-_nl2a_wrapper_tool = local_mcp_service.tool(
+local_mcp_service.tool(
     _nl2a_wrapper,
     name=NL2A_WRAPPER_NAME,
-    description=(
-        "Build one NL2Agent output from subtype-specific parameters. Always pass "
-        "`subtype`. For `local_mcp_recommendation`, also pass `search_result` and "
-        "`selected_tool_ids`. For `agent_draft`, pass the agent draft fields. Call "
-        "the tool as `result = nl2a_wrapper(...)`, then use `print(result)`."
-    ),
-    meta={"nexent_internal": True},
+    description=NL2A_WRAPPER_DESCRIPTION,
+    meta=NL2AGENT_MCP_TOOL_META,
 )
-
-
-def get_nl2agent_mcp_tool_descriptions() -> dict[str, str]:
-    """Return descriptions from the registered NL2Agent MCP tools."""
-    return {
-        SEARCH_INSTALLED_MCP_TOOLS_NAME: _search_installed_mcp_tool.description,
-        NL2A_WRAPPER_NAME: _nl2a_wrapper_tool.description,
-    }
 
 
 @local_mcp_service.tool(name="test_tool_name",
