+++ test/backend/services/test_mcp_internal_tool_search.py
@@ -7,18 +7,21 @@ from fastmcp import FastMCP
 from mcp.types import Tool
 from pydantic import ValidationError
 
-from agents.nl2agent_agent import Nl2aAgentDraftInput, Nl2aLocalMcpRecommendationInput
 import services.nl2agent_service as nl2agent_service
 import services.tool_configuration_service as tool_configuration_service
 import tool_collection.mcp.nl2agent_mcp_tools as nl2agent_mcp_tools_module
 from services.tool_configuration_service import get_tool_from_remote_mcp_server
 from tool_collection.mcp.local_mcp_service import (
     LOCAL_MCP_TOOL_NAME_OVERRIDES,
-    get_nl2agent_mcp_tool_descriptions,
     local_mcp_service,
 )
 from tool_collection.mcp.nl2agent_mcp_tools import (
+    NL2AGENT_MCP_TOOL_META,
+    NL2A_WRAPPER_DESCRIPTION,
     NL2A_WRAPPER_NAME,
+    Nl2aAgentDraftInput,
+    Nl2aLocalMcpRecommendationInput,
+    SEARCH_INSTALLED_MCP_TOOLS_DESCRIPTION,
     SEARCH_INSTALLED_MCP_TOOLS_NAME,
     nl2a_wrapper,
     search_installed_mcp_tools,
@@ -490,13 +493,12 @@ async def test_mcp_search_registration_has_stable_name_schema_and_marker():
     assert tool.parameters["properties"]["keywords"]["type"] == "array"
     assert tool.parameters["properties"]["keywords"]["items"]["type"] == "string"
     assert tool.parameters["required"] == ["keywords"]
-    assert tool.meta["nexent_internal"] is True
+    assert tool.meta == NL2AGENT_MCP_TOOL_META
     assert "print(result)" in tool.description
-    assert get_nl2agent_mcp_tool_descriptions() == {
-        SEARCH_INSTALLED_MCP_TOOLS_NAME: tool.description,
-        NL2A_WRAPPER_NAME: wrapper_tool.description,
-    }
+    assert tool.description == SEARCH_INSTALLED_MCP_TOOLS_DESCRIPTION
     assert wrapper_tool.name == NL2A_WRAPPER_NAME
+    assert wrapper_tool.description == NL2A_WRAPPER_DESCRIPTION
+    assert wrapper_tool.meta == NL2AGENT_MCP_TOOL_META
     assert wrapper_tool.parameters["required"] == ["subtype"]
     assert set(wrapper_tool.parameters["properties"]) == {
         "subtype",
